pyproject.toml
[project]
name = "py-crypter"
version = "2.1.3"
description = "NStor PY_Crypter AES-256"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "pycryptodome>=3.23.0",
]
